Redux file reorganization
