+++ spotifyApi/shuffle/front/src/appState/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import counterReducer from '../features/counter/counterSlice';
+import counterReducer from './slices/counter';
 
 export const store = configureStore({
     reducer: {
