+++ spotifyApi/shuffle/front/tsconfig.json
@@ -22,6 +22,6 @@
     "noFallthroughCasesInSwitch": true
   },
   "include": ["src"],
-  "exclude": ["src/features/dogs"],
+  "exclude": ["src/appState/dogs.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
